Separate out scoring related tests
